Support 0-latency gates. (#1)
Adds support for 0-latency 1 and 2 qubit gates, with the exception of swap gates, which must still be of a non-zero duration. This change was made to support instructions like RZ on IBM hardware, which is considered to execute instantaneously.

Prior to this change, 0-latency gates caused libtoqm to crash.

To support such gates, Node::scheduleGate has been changed to immediately schedule any subsequent 0-latency gates that become unblocked (all qubit data dependencies satisfied) as a result of scheduling the given gate. This is recursive for consecutive 0-latency gates.