Fix 0-latency scheduling. (#4)
Previously, it was possible for two non-zero latency gates to clobber each other at the same cycle if a zero latency gate was scheduled between them.

This happened because scheduling a zero-latency gate on the same qubits as a previously scheduled non-zero-latency gate updates `Environment::lastGate`, which is used by `Node::busyCycles` to report for how long a qubit will be busy. To fix this, a separate `Environment::lastNonZeroLatencyGate` is maintained for the purpose of calculating busy cycles.